改sm4 加密算法

+++ school-common/src/main/java/com/ruoyi/common/utils/SM/SM4_Context.java
@@ -1,15 +1,17 @@
 package com.ruoyi.common.utils.SM;
-public class SM4_Context {
-	public int mode;  
+
+public class SM4_Context
+{
+    public int mode;
     
-    public long[] sk;  
-      
-    public boolean isPadding;  
-  
-    public SM4_Context()   
-    {  
-        this.mode = 1;  
-        this.isPadding = true;  
-        this.sk = new long[32];  
-    }  
+    public int[] sk;
+    
+    public boolean isPadding;
+
+    public SM4_Context() 
+    {
+        this.mode = 1;
+        this.isPadding = true;
+        this.sk = new int[32];
+    }
 }
